fixed lint

+++ contracts/Rewards/RewardsDistributor.sol
@@ -82,7 +82,7 @@ contract RewardsDistributor is ExponentialNoError, Ownable2StepUpgradeable {
     event RewardTokenBorrowIndexUpdated(address vToken, Exp marketBorrowIndex);
 
     /// @notice Emitted when a reward for contributor is updated
-    event ContributorRewardsUpdated(address contributor, uint rewardAccrued);
+    event ContributorRewardsUpdated(address contributor, uint256 rewardAccrued);
 
     /// @notice The REWARD TOKEN borrow index for each market for each borrower as of the last time they accrued REWARD TOKEN
     mapping(address => mapping(address => uint256)) public rewardTokenBorrowerIndex;
